+++ CHANGES.md
@@ -28,6 +28,7 @@ Changes in 12.9
 - Remove duplicate GET params from form URL (start & end date, others) in Percent.php, CategoryBreakdown.php, CategoryBreakdownTime.php, StudentFieldBreakdown.php, TransactionsReport.php, AddDrop.php, Calendar.php, StudentBreakdown.php & AddDropBreakdownTime.php
 - Security fix #375 Authenticated Time-based SQL Injection (non-student roles) in StudentSumary.php, thanks to @Cresp0
 - Security fix #376 IDOR: Unauthorized Cross-School Data Assignment in AddAbsences.php, AddActivity.php, Requests.php, MassAssignFees.php, MassAssignPayments.php, AssignOtherInfo.php & AddStudents.php, thanks to @Cresp0
+- HTML display checkbox before its label in MailingLabels.php, ProgressReports.php & PrintStudentInfo.php
 
 Changes in 12.8.1
 -----------------